Remove css & scss and consolidate base styling into base.css
All other styling now needs to be done on a component level. This requires a lot of fixing as there is styling within the SCSS that needs to be incorporated into some components. The layout needs redoing but this is a chance to make the site a bit more responsive and mobile friendly.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>